feature: Complete CallvanNotificationsViewModel with UseCase
